Changing input event generation, remove unnecessary batch testcases, added multithreaded evalauation and efficient predicate selection

+++ src/state/EndState.java
@@ -49,7 +49,7 @@ public class EndState implements State{
 	}
 
 	@Override
-	public void pumpHeartbeat(long heartbeat) {
+	public void pumpHeartbeat(double heartbeat) {
 	}
 
 }
